requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
langchain==1.1.0
langchain-azure-ai==1.0.3
langchain-community==0.4.1
python-dotenv==1.2.1
pydantic==2.12.5
httpx==0.28.1
playwright==1.56.0
beautifulsoup4==4.14.2
pypdf2==3.0.1
pandas==2.3.3
pillow==12.0.0
aiofiles==25.1.0
requests==2.32.5
lxml==6.0.2